refactor createconfig prompts
